requirements.txt
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pip
python-dotenv
ruff
scikit-learn
tqdm
typer
openpyxl
streamlit
jira
psycopg2-binary
pgvector
langchain
langchain-ollama
ollama
langchain_openai
langchain_community
langchain_google
-e .
